requirements.txt
streamlit 
openai 
python-dotenv 
pymupdf 
python-docx 
